// packages/rma-server/src/purchase-order/entity/purchase-order/purchase-order-item.interface.ts
export interface PurchaseOrderItem {
  name: string;
  owner: string;
  creation: string;
  modified: string;
  modified_by: string;
  parent: string;
  parentfield: string;
  parenttype: string;
  idx: number;
  docstatus: number;
  item_code: string;
  item_name: string;
  schedule_date: string;
  description: string;
  is_nil_exempt: number;
  is_non_gst: number;
  image: string;
  qty: number;
  stock_uom: string;
  uom: string;
  conversion_factor: number;
  price_list_rate: number;
  discount_percentage: number;
  discount_amount: number;
  last_purchase_rate: number;
  base_price_list_rate: number;
  rate: number;
  amount: number;
  base_rate: number;
  base_amount: number;
  is_free_item: number;
  is_fixed_asset: number;
  net_rate: number;
  net_amount: number;
  base_net_rate: number;
  base_net_amount: number;
  billed_amt: number;
  weight_per_unit: number;
  total_weight: number;
  warehouse: string;
  delivered_by_supplier: number;
  against_blanket_order: number;
  blanket_order_rate: number;
  item_group: string;
  include_exploded_items: number;
  stock_qty: number;
  received_qty: number;
  returned_qty: number;
  page_break: number;
  item_tax_rate: string;
  doctype: string;
}